Made Ninja to have minimum distance for shooting.   If it is too close it will walk away and face back to the target after walking and then patrol and shoot again.
