+++ Assets/Game/ScriptableObjects/States/ChaseState.asset
@@ -25,6 +25,6 @@ MonoBehaviour:
     trueState: {fileID: 11400000, guid: 0e0d8176683f9442488f7f26685a0030, type: 2}
     falseState: {fileID: 11400000, guid: 061da9034edb84c1287ca2f2e4112e91, type: 2}
   - decision: {fileID: 11400000, guid: 89910f032d9454c1f984022b8e76ba44, type: 2}
-    trueState: {fileID: 11400000, guid: f4afac98793f24fe2b21ae96e32fa0df, type: 2}
+    trueState: {fileID: 11400000, guid: 98568f79680a842218fa6c7f95854fc9, type: 2}
     falseState: {fileID: 11400000, guid: 061da9034edb84c1287ca2f2e4112e91, type: 2}
   sceneGizmoColor: {r: 1, g: 0.98470926, b: 0, a: 1}
