--- Assets/Game/ScriptableObjects/States/FaceToTargetState.asset
@@ -1,21 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 8dc87918e45914cb8b114d4e5536adc2, type: 3}
-  m_Name: FaceToTargetState
-  m_EditorClassIdentifier: 
-  actions:
-  - {fileID: 11400000, guid: 11af2852362bb45c4a2a91531db6ae27, type: 2}
-  transitions:
-  - decision: {fileID: 11400000, guid: 93f9142651e67461487393ca61a88b97, type: 2}
-    trueState: {fileID: 11400000, guid: 52d49f92640914388b50a008787f2fd6, type: 2}
-    falseState: {fileID: 11400000, guid: 52d49f92640914388b50a008787f2fd6, type: 2}
-  sceneGizmoColor: {r: 0.29803923, g: 0.29803923, b: 0.29803923, a: 1}
