--- Assets/Game/ScriptableObjects/States/FleeRemainState.asset
@@ -1,21 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 8dc87918e45914cb8b114d4e5536adc2, type: 3}
-  m_Name: FleeRemainState
-  m_EditorClassIdentifier: 
-  actions:
-  - {fileID: 11400000, guid: dd22d4757ff2f46f39fd0c07912cd694, type: 2}
-  transitions:
-  - decision: {fileID: 11400000, guid: b7993c69252dc4815ae47f4d4a7f4a2e, type: 2}
-    trueState: {fileID: 11400000, guid: 061da9034edb84c1287ca2f2e4112e91, type: 2}
-    falseState: {fileID: 11400000, guid: 4fc62cdefdcbb45a98f2bedcf820ec83, type: 2}
-  sceneGizmoColor: {r: 0.7924023, g: 0.022472408, b: 0.9528302, a: 1}
