--- Assets/Game/ScriptableObjects/States/FleeRemainState.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 061c68c3b714049278af78da13ad23d1
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
