+++ Assets/Game/ScriptableObjects/States/FleeState.asset
@@ -9,13 +9,12 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 8dc87918e45914cb8b114d4e5536adc2, type: 3}
+  m_Script: {fileID: 11500000, guid: a0f5122133c64c02861c7c1243c71752, type: 3}
   m_Name: FleeState
   m_EditorClassIdentifier: 
-  actions:
-  - {fileID: 11400000, guid: e6e324a3427d24fd38b45601e9592855, type: 2}
+  actions: []
   transitions:
   - decision: {fileID: 11400000, guid: b7993c69252dc4815ae47f4d4a7f4a2e, type: 2}
-    trueState: {fileID: 11400000, guid: 061c68c3b714049278af78da13ad23d1, type: 2}
+    trueState: {fileID: 11400000, guid: 061da9034edb84c1287ca2f2e4112e91, type: 2}
     falseState: {fileID: 11400000, guid: 52d49f92640914388b50a008787f2fd6, type: 2}
-  sceneGizmoColor: {r: 0.54901963, g: 0.36532933, b: 0, a: 1}
+  sceneGizmoColor: {r: 1, g: 0, b: 0.97048855, a: 1}
