+++ Assets/Game/ScriptableObjects/States/Thanatos.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 4fc62cdefdcbb45a98f2bedcf820ec83
-NativeFormatImporter:
+guid: 9485a87cdeec14e74bbb73eeea0897ba
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
