+++ Assets/Game/Scripts/Spawn.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 82a6b30b104c441119c6450a15932891
+guid: bf908998a5364fff96b60c3e15ae7e36
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
