--- Assets/Game/Scripts/StateMachine/Actions/DoNothingAction.cs
@@ -1,11 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-[CreateAssetMenu (menuName = "StateMachine/Actions/Do Nothing")]
-public class DoNothingAction : Action 
-{
-  public override void Act (AI ai)
-  {
-  }
-}
