--- Assets/Game/Scripts/StateMachine/Actions/DoNothingAction.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: b7ae396117b640079eaf3c1374fc5a6d
-timeCreated: 1613343124
