--- Assets/Game/Scripts/StateMachine/Actions/FaceToTargetAction.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: abfa02d78f1f4361884a6cc78cbe411d
-timeCreated: 1613351558
