--- Assets/Game/Scripts/StateMachine/Actions/FleeAction.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 06b6c593fc5e40d3aa8179ca3d12bc5d
-timeCreated: 1613272896
