--- Assets/StreamingAssets.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ca5b67bc149584b57a28418d3182e1b4
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
